+++ app/javascript/components/index.js
@@ -1,8 +1,8 @@
 // javascript/components/index.js
 export { default as AnimatedTitle } from './AnimatedTitle'
-export { default as Pantry } from './Pantry'
-export { default as GroceryItem } from './GroceryItem'
-export { default as RecipeBox } from './RecipeBox'
-export { default as RecipeDetail } from './RecipeDetail'
+export { default as Pantry } from './grocery/Pantry'
+export { default as GroceryItem } from './grocery/GroceryItem'
+export { default as RecipeBox } from './recipe/RecipeBox'
+export { default as RecipeDetail } from './recipe/RecipeDetail'
 
 // Add more components here
